feat: onboarding animation add, aggreagate to one page
